Refactor translate/filters/checks.py into a package
