+++ tests/translate/lang/test_ro.py
@@ -1,8 +1,7 @@
 import string
 
+from tests.translate.filters.checks.helpers import fails, passes
 from translate.lang.ro import RomanianChecker
-
-from ..filters.test_checks import fails, passes
 
 
 def test_cedillas() -> None:
