+++ translate/filters/checks/cclicense.py
@@ -1,0 +1,37 @@
+#
+# Copyright 2004-2011 Zuza Software Foundation
+# 2013, 2016 F Wolff
+#
+# This file is part of translate.
+#
+# translate is free software; you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation; either version 3 of the License, or
+# (at your option) any later version.
+#
+# translate is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, see <https://www.gnu.org/licenses/>.
+
+"""Checks specific to Creative Commons licenses."""
+
+from translate.filters.checks.config import CheckerConfig
+from translate.filters.checks.standard import StandardChecker
+
+cclicenseconfig = CheckerConfig(varmatches=[("@", "@")])
+
+
+class CCLicenseChecker(StandardChecker):
+    def __init__(self, **kwargs) -> None:
+        checkerconfig = kwargs.get("checkerconfig")
+
+        if checkerconfig is None:
+            checkerconfig = CheckerConfig()
+            kwargs["checkerconfig"] = checkerconfig
+
+        checkerconfig.update(cclicenseconfig)
+        super().__init__(**kwargs)
