+++ translate/filters/checks/standard.py
@@ -17,30 +17,15 @@
 # You should have received a copy of the GNU General Public License
 # along with this program; if not, see <https://www.gnu.org/licenses/>.
 
-"""
-This is a set of validation checks that can be performed on translation
-units.
+"""The standard set of checks for (source, target) translation pairs."""
 
-Derivatives of UnitChecker (like StandardUnitChecker) check translation units,
-and derivatives of TranslationChecker (like StandardChecker) check
-(source, target) translation pairs.
-
-When adding a new test here, please document and explain their behaviour on the
-:doc:`pofilter tests </commands/pofilter_tests>` page.
-"""
-
-from __future__ import annotations
-
-import logging
 import re
-import string
 
 from translate.filters import decoration, helpers, prefilters, spelling
+from translate.filters.checks.checker import TranslationChecker
+from translate.filters.checks.exceptions import FilterFailure, SeriousFilterFailure
+from translate.filters.checks.tags import intuplelist, tag_re, tagproperties
 from translate.filters.decorators import cosmetic, critical, extraction, functional
-from translate.lang import data, factory
-from translate.storage import base
-
-logger = logging.getLogger(__name__)
 
 # These are some regular expressions that are compiled for use in some tests
 
@@ -67,630 +52,7 @@ printf_pat = re.compile(
     re.VERBOSE,
 )
 
-# The name of the XML tag
-tagname_re = re.compile(r"<[\s]*([\w\/]*).*?(/)?[\s]*>", re.DOTALL)
-
-# We allow escaped quotes, probably for old escaping style of OOo helpcontent
-# TODO: remove escaped strings once usage is audited
-property_re = re.compile(r""" (\w*)=((\\?".*?\\?")|(\\?'.*?\\?'))""")
-
-# The whole tag
-tag_re = re.compile(r"<[^>]+>")
-
 anonvar_re = re.compile(r"^{[0-9]*}$")
-
-gconf_attribute_re = re.compile(r'"[a-z_]+?"')
-
-# XML/HTML tags in LibreOffice help and readme, exclude short tags
-lo_tag_re = re.compile(r"""</?(?P<tag>[a-z][a-z_-]+)(?: +[a-z]+="[^"]+")* */?>""")
-lo_emptytags = frozenset(["br", "embed", "embedvar", "object", "help-id-missing"])
-TagProperty = tuple[str | None, str | None, str | None]
-
-
-def tagname(string: str) -> str:
-    """Returns the name of the XML/HTML tag in string."""
-    tagname_match = tagname_re.match(string)
-    assert tagname_match is not None, f"Expected tag in string: {string}"
-    # Extract the tag name (group 1) and optional slash (group 2)
-    group1 = tagname_match.group(1) or ""
-    group2 = tagname_match.group(2) or ""
-    return group1 + group2
-
-
-def intuplelist(pair: TagProperty, patterns: list[TagProperty]) -> TagProperty:
-    """
-    Tests to see if pair == (a,b,c) is in list, but handles None entries in
-    list as wildcards (only allowed in positions "a" and "c"). We take a
-    shortcut by only considering "c" if "b" has already matched.
-    """
-    a, b, c = pair
-
-    if (b, c) == (None, None):
-        # This is a tagname
-        return pair
-
-    for pattern in patterns:
-        x, y, z = pattern
-
-        if (x, y) in {(a, b), (None, b)} and z in {None, c}:
-            return pattern
-
-    return pair
-
-
-def tagproperties(matches: list[str], ignore: list[TagProperty]) -> list[TagProperty]:
-    """
-    Returns all the properties in the XML/HTML tag string as (tagname,
-    propertyname, propertyvalue), but ignore those combinations specified in
-    ignore.
-    """
-    properties: list[TagProperty] = []
-
-    for match in matches:
-        tag = tagname(match)
-        properties += [(tag, None, None)]
-        # Now we isolate the attribute pairs.
-        pairs = property_re.findall(match)
-
-        for property, value, _a, _b in pairs:
-            # Strip the quotes:
-            value = value[1:-1]
-
-            canignore = False
-
-            if (tag, property, value) in ignore or intuplelist(
-                (tag, property, value), ignore
-            ) != (tag, property, value):
-                canignore = True
-                break
-
-            if not canignore:
-                properties += [(tag, property, value)]
-
-    return properties
-
-
-class FilterFailure(Exception):
-    """
-    This exception signals that a Filter didn't pass, and gives an
-    explanation or a comment.
-    """
-
-    def __init__(self, messages) -> None:
-        if not isinstance(messages, list):
-            messages = [messages]
-
-        assert isinstance(messages[0], str)  # Assumption: all of same type
-
-        self.messages = messages
-
-    def __str__(self) -> str:
-        return ", ".join(self.messages)
-
-
-class SeriousFilterFailure(FilterFailure):
-    """
-    This exception signals that a Filter didn't pass, and the bad
-    translation might break an application (so the string will be marked
-    fuzzy).
-    """
-
-
-# (tag, attribute, value) specifies a certain attribute which can be changed/
-# ignored if it exists inside tag. In the case where there is a third element
-# in the tuple, it indicates a property value that can be ignored if present
-# (like defaults, for example)
-# If a certain item is None, it indicates that it is relevant for all values of
-# the property/tag that is specified as None. A non-None value of "value"
-# indicates that the value of the attribute must be taken into account.
-common_ignoretags = [(None, "xml-lang", None)]
-common_canchangetags = [
-    ("img", "alt", None),
-    (None, "title", None),
-    (None, "dir", None),
-    (None, "lang", None),
-]
-# Actually the title tag is allowed on many tags in HTML (but probably not all)
-
-
-class CheckerConfig:
-    """Object representing the configuration of a checker."""
-
-    def __init__(
-        self,
-        targetlanguage=None,
-        accelmarkers=None,
-        varmatches=None,
-        notranslatewords=None,
-        musttranslatewords=None,
-        validchars=None,
-        punctuation=None,
-        endpunctuation=None,
-        ignoretags=None,
-        canchangetags=None,
-        criticaltests=None,
-        credit_sources=None,
-    ) -> None:
-        # Init lists
-        self.accelmarkers = self._init_list(accelmarkers)
-        self.varmatches = self._init_list(varmatches)
-        self.criticaltests = self._init_list(criticaltests)
-        self.credit_sources = self._init_list(credit_sources)
-
-        # Lang data
-        self.updatetargetlanguage(targetlanguage)
-        self.sourcelang = factory.getlanguage("en")
-
-        # Inits with default values
-        self.punctuation = self._init_default(
-            data.normalize(punctuation), self.lang.punctuation
-        )
-        self.endpunctuation = self._init_default(
-            data.normalize(endpunctuation), self.lang.sentenceend
-        )
-        self.ignoretags = self._init_default(ignoretags, common_ignoretags)
-        self.canchangetags = self._init_default(canchangetags, common_canchangetags)
-
-        # Other data
-        # TODO: allow user configuration of untranslatable words
-        self.notranslatewords = dict.fromkeys(
-            [data.normalize(key) for key in self._init_list(notranslatewords)]
-        )
-        self.musttranslatewords = dict.fromkeys(
-            [data.normalize(key) for key in self._init_list(musttranslatewords)]
-        )
-        validchars = data.normalize(validchars)
-        self.validcharsmap = {}
-        self.updatevalidchars(validchars)
-
-    @staticmethod
-    def _init_list(list: list | None) -> list:
-        """
-        Initialise configuration parameters that are lists.
-
-        :param list: None (we'll initialise a blank list) or a list parameter
-        """
-        if list is None:
-            list = []
-
-        return list
-
-    @staticmethod
-    def _init_default(param, default):
-        """
-        Initialise parameters that can have default options.
-
-        :param param: the user supplied parameter value
-        :param default: default values when param is not specified
-        :return: the parameter as specified by the user of the default settings
-        """
-        if param is None:
-            return default
-
-        return param
-
-    def update(self, otherconfig) -> None:
-        """Combines the info in ``otherconfig`` into this config object."""
-        self.targetlanguage = otherconfig.targetlanguage or self.targetlanguage
-        self.updatetargetlanguage(self.targetlanguage)
-        self.accelmarkers.extend(
-            [c for c in otherconfig.accelmarkers if c not in self.accelmarkers]
-        )
-        self.varmatches.extend(otherconfig.varmatches)
-        self.notranslatewords.update(otherconfig.notranslatewords)
-        self.musttranslatewords.update(otherconfig.musttranslatewords)
-        self.validcharsmap.update(otherconfig.validcharsmap)
-        self.punctuation += otherconfig.punctuation
-        self.endpunctuation += otherconfig.endpunctuation
-        # TODO: consider also updating in the following cases:
-        self.ignoretags = otherconfig.ignoretags
-        self.canchangetags = otherconfig.canchangetags
-        self.criticaltests.extend(otherconfig.criticaltests)
-        self.credit_sources = otherconfig.credit_sources
-
-    def updatevalidchars(self, validchars) -> None:
-        """Updates the map that eliminates valid characters."""
-        if validchars is None:
-            return
-
-        validcharsmap = {
-            ord(validchar): None for validchar in data.normalize(validchars)
-        }
-        self.validcharsmap.update(validcharsmap)
-
-    def updatetargetlanguage(self, langcode) -> None:
-        """
-        Updates the target language in the config to the given target
-        language and sets its script.
-        """
-        self.targetlanguage = langcode
-        self.lang = factory.getlanguage(langcode)
-        self.language_script = ""
-
-        for script, langs in data.scripts.items():
-            if langcode in langs or data.simplercode(langcode) in langs:
-                self.language_script = script
-                break
-
-
-def cache_results(f):
-    def cached_f(self, param1):
-        key = (f.__name__, param1)
-        res_cache = self.results_cache
-
-        if key in res_cache:
-            return res_cache[key]
-        value = f(self, param1)
-        res_cache[key] = value
-        return value
-
-    return cached_f
-
-
-class UnitChecker:
-    """
-    Parent Checker class which does the checking based on functions
-    available in derived classes.
-    """
-
-    preconditions = {}
-
-    def __init__(
-        self,
-        checkerconfig=None,
-        excludefilters=None,
-        limitfilters=None,
-        errorhandler=None,
-    ) -> None:
-        self.errorhandler = errorhandler
-
-        #: Categories where each checking function falls into
-        #: Function names are used as keys, categories are the values
-        self.categories = {}
-
-        if checkerconfig is None:
-            self.setconfig(CheckerConfig())
-        else:
-            self.setconfig(checkerconfig)
-
-        # Exclude functions defined in UnitChecker from being treated as tests.
-        self.helperfunctions = {}
-
-        for functionname in dir(UnitChecker):
-            function = getattr(self, functionname)
-
-            if callable(function):
-                self.helperfunctions[functionname] = function
-
-        self.defaultfilters = self.getfilters(excludefilters, limitfilters)
-        self.results_cache = {}
-
-    def getfilters(self, excludefilters=None, limitfilters=None):
-        """
-        Returns dictionary of available filters, including/excluding those
-        in the given lists.
-        """
-        filters = {}
-
-        if limitfilters is None:
-            # use everything available unless instructed
-            limitfilters = dir(self)
-
-        if excludefilters is None:
-            excludefilters = {}
-
-        for functionname in limitfilters:
-            if functionname in excludefilters:
-                continue
-
-            if functionname in self.helperfunctions:
-                continue
-
-            if functionname == "errorhandler":
-                continue
-
-            filterfunction = getattr(self, functionname, None)
-            if not callable(filterfunction):
-                continue
-
-            filters[functionname] = filterfunction
-
-        return filters
-
-    def setconfig(self, config) -> None:
-        """Sets the accelerator list."""
-        self.config = config
-        self.accfilters = [
-            prefilters.filteraccelerators(accelmarker)
-            for accelmarker in self.config.accelmarkers
-        ]
-        self.varfilters = [
-            prefilters.filtervariables(startmatch, endmatch, prefilters.varname)
-            for startmatch, endmatch in self.config.varmatches
-        ]
-        self.removevarfilter = [
-            prefilters.filtervariables(startmatch, endmatch, prefilters.varnone)
-            for startmatch, endmatch in self.config.varmatches
-        ]
-
-    def setsuggestionstore(self, store) -> None:
-        """
-        Sets the filename that a checker should use for evaluating
-        suggestions.
-        """
-        self.suggestion_store = store
-
-        if self.suggestion_store:
-            self.suggestion_store.require_index()
-
-    @cache_results
-    def filtervariables(self, str1):
-        """Filter out variables from ``str1``."""
-        return helpers.multifilter(str1, self.varfilters)
-
-    @cache_results
-    def removevariables(self, str1):
-        """Remove variables from ``str1``."""
-        return helpers.multifilter(str1, self.removevarfilter)
-
-    @cache_results
-    def filteraccelerators(self, str1):
-        """Filter out accelerators from ``str1``."""
-        return helpers.multifilter(str1, self.accfilters, None)
-
-    def filteraccelerators_by_list(self, str1, acceptlist=None):
-        """Filter out accelerators from ``str1``."""
-        return helpers.multifilter(str1, self.accfilters, acceptlist)
-
-    @cache_results
-    def filterwordswithpunctuation(self, str1):
-        """
-        Replaces words with punctuation with their unpunctuated
-        equivalents.
-        """
-        return prefilters.filterwordswithpunctuation(str1)
-
-    @cache_results
-    def filterxml(self, str1):
-        """Filter out XML from the string so only text remains."""
-        return tag_re.sub("", str1)
-
-    @staticmethod
-    def run_test(test, unit):
-        """
-        Runs the given test on the given unit.
-
-        Note that this can raise a :exc:`FilterFailure` as part of normal operation.
-        """
-        return test(unit)
-
-    @property
-    def checker_name(self):
-        """Extract checker name, for example 'mozilla' from MozillaChecker."""
-        return str(self.__class__.__name__).lower()[: -len("checker")]
-
-    def get_ignored_filters(self):
-        """Return checker's additional filters for current language."""
-        return list(
-            set(
-                self.config.lang.ignoretests.get(self.checker_name, [])
-                + self.config.lang.ignoretests.get("all", [])
-            )
-        )
-
-    def run_filters(self, unit, categorised: bool = False) -> dict[str, dict]:
-        """
-        Run all the tests in this suite.
-
-        :return: Content of the dictionary is as follows::
-
-           {'testname': { 'message': message_or_exception, 'category': failure_category } }
-        """
-        self.results_cache = {}
-        failures = {}
-        ignores = self.get_ignored_filters()
-        functionnames = self.defaultfilters.keys()
-        priorityfunctionnames = self.preconditions.keys()
-        otherfunctionnames = filter(
-            lambda functionname: functionname not in self.preconditions, functionnames
-        )
-
-        for functionname in list(priorityfunctionnames) + list(otherfunctionnames):
-            if functionname in ignores:
-                continue
-
-            filterfunction = getattr(self, functionname, None)
-
-            # This filterfunction may only be defined on another checker if
-            # using TeeChecker
-            if filterfunction is None:
-                continue
-
-            filtermessage = ""
-
-            try:
-                filterresult = self.run_test(filterfunction, unit)
-            except FilterFailure as e:
-                filterresult = False
-                filtermessage = str(e)
-            except Exception as e:
-                if self.errorhandler is None:
-                    raise ValueError(
-                        f"error in filter {functionname}: {unit.source!r}, {unit.target!r}, {e}"
-                    ) from e
-                filterresult = self.errorhandler(
-                    functionname, unit.source, unit.target, e
-                )
-            if not filterresult:
-                if not filtermessage:
-                    # Should be quite rare
-                    # pylint: disable-next=import-outside-toplevel
-                    import pydoc  # ruff:ignore[import-outside-top-level]
-
-                    # Strip out unnecessary whitespace from docstring
-                    filtermessage = pydoc.getdoc(filterfunction)
-                # We test some preconditions that aren't actually a cause for
-                # failure
-                if functionname in self.defaultfilters:
-                    failures[functionname] = {
-                        "message": filtermessage,
-                        "category": self.categories[functionname],
-                    }
-
-                if functionname in self.preconditions:
-                    for ignoredfunctionname in self.preconditions[functionname]:
-                        ignores.append(ignoredfunctionname)
-
-        self.results_cache = {}
-
-        if not categorised:
-            for name, info in failures.items():
-                failures[name] = info["message"]
-        return failures
-
-
-class TranslationChecker(UnitChecker):
-    """
-    A checker that passes source and target strings to the checks, not the
-    whole unit.
-
-    This provides some speedup and simplifies testing.
-    """
-
-    def __init__(
-        self,
-        checkerconfig=None,
-        excludefilters=None,
-        limitfilters=None,
-        errorhandler=None,
-    ) -> None:
-        super().__init__(checkerconfig, excludefilters, limitfilters, errorhandler)
-
-        self.locations = []
-
-    def run_test(self, test, unit):
-        """
-        Runs the given test on the given unit.
-
-        Note that this can raise a :exc:`FilterFailure` as part of normal
-        operation.
-        """
-        if self.hasplural:
-            filtermessages = []
-            filterresult = True
-
-            for pluralform in unit.target.strings:
-                try:
-                    if not test(self.str1, str(pluralform)):
-                        filterresult = False
-                except FilterFailure as e:
-                    filterresult = False
-                    filtermessages.extend(e.messages)
-
-            if not filterresult and filtermessages:
-                raise FilterFailure(filtermessages)
-            return filterresult
-        return test(self.str1, self.str2)
-
-    def run_filters(self, unit, categorised=False):
-        """
-        Do some optimisation by caching some data of the unit for the
-        benefit of :meth:`~TranslationChecker.run_test`.
-        """
-        self.str1 = data.normalize(unit.source) or ""
-        self.str2 = data.normalize(unit.target) or ""
-        self.hasplural = unit.hasplural()
-        self.locations = unit.getlocations()
-
-        return super().run_filters(unit, categorised)
-
-
-class TeeChecker:
-    """A Checker that controls multiple checkers."""
-
-    #: Categories where each checking function falls into
-    #: Function names are used as keys, categories are the values
-    categories = {}
-
-    def __init__(
-        self,
-        checkerconfig=None,
-        excludefilters=None,
-        limitfilters=None,
-        checkerclasses=None,
-        errorhandler=None,
-        languagecode=None,
-    ) -> None:
-        """Construct a TeeChecker from the given checkers."""
-        self.limitfilters = limitfilters
-
-        if checkerclasses is None:
-            checkerclasses = [StandardChecker]
-
-        self.checkers = [
-            checkerclass(
-                checkerconfig=checkerconfig,
-                excludefilters=excludefilters,
-                limitfilters=limitfilters,
-                errorhandler=errorhandler,
-            )
-            for checkerclass in checkerclasses
-        ]
-
-        if languagecode:
-            for checker in self.checkers:
-                checker.config.updatetargetlanguage(languagecode)
-
-            # Let's hook up the language specific checker
-            lang_checker = self.checkers[0].config.lang.checker
-
-            if lang_checker:
-                self.checkers.append(lang_checker)
-
-        self.combinedfilters = self.getfilters(excludefilters, limitfilters)
-        self.config = checkerconfig or self.checkers[0].config
-
-    def getfilters(self, excludefilters=None, limitfilters=None):
-        """
-        Returns a dictionary of available filters, including/excluding
-        those in the given lists.
-        """
-        if excludefilters is None:
-            excludefilters = {}
-
-        filterslist = [
-            checker.getfilters(excludefilters, limitfilters)
-            for checker in self.checkers
-        ]
-        self.combinedfilters = {}
-
-        for filters in filterslist:
-            self.combinedfilters.update(filters)
-
-        # TODO: move this somewhere more sensible (a checkfilters method?)
-        if limitfilters is not None:
-            for filtername in limitfilters:
-                if filtername not in self.combinedfilters:
-                    logger.warning("could not find filter %s", filtername)
-
-        return self.combinedfilters
-
-    def run_filters(self, unit, categorised=False):
-        """Run all the tests in the checker's suites."""
-        failures = {}
-
-        for checker in self.checkers:
-            failures.update(checker.run_filters(unit, categorised))
-
-        return failures
-
-    def setsuggestionstore(self, store) -> None:
-        """
-        Sets the filename that a checker should use for evaluating
-        suggestions.
-        """
-        for checker in self.checkers:
-            checker.setsuggestionstore(store)
 
 
 class StandardChecker(TranslationChecker):
@@ -2216,695 +1578,3 @@ class StandardChecker(TranslationChecker):
             "validxml",
         ),
     }
-
-
-# code to actually run the tests (use unittest?)
-
-
-openofficeconfig = CheckerConfig(
-    accelmarkers=["~"],
-    varmatches=[
-        ("&", ";"),
-        ("%", "%"),
-        ("%", None),
-        ("%", 0),
-        ("$(", ")"),
-        ("$", "$"),
-        ("${", "}"),
-        ("#", "#"),
-        ("#", 1),
-        ("#", 0),
-        ("($", ")"),
-        ("$[", "]"),
-        ("[", "]"),
-        ("@", "@"),
-        ("$", None),
-    ],
-    ignoretags=[
-        ("alt", "xml-lang", None),
-        ("ahelp", "visibility", "visible"),
-        ("img", "width", None),
-        ("img", "height", None),
-    ],
-    canchangetags=[("link", "name", None)],
-)
-
-
-class OpenOfficeChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(openofficeconfig)
-        super().__init__(**kwargs)
-
-
-libreofficeconfig = CheckerConfig(
-    accelmarkers=["~"],
-    varmatches=[
-        ("&", ";"),
-        ("%", "%"),
-        ("%", None),
-        ("%", 0),
-        ("$(", ")"),
-        ("$", "$"),
-        ("${", "}"),
-        ("#", "#"),
-        ("#", 1),
-        ("#", 0),
-        ("($", ")"),
-        ("$[", "]"),
-        ("[", "]"),
-        ("@", "@"),
-        ("$", None),
-    ],
-    ignoretags=[
-        ("alt", "xml-lang", None),
-        ("ahelp", "visibility", "visible"),
-        ("img", "width", None),
-        ("img", "height", None),
-    ],
-    canchangetags=[("link", "name", None)],
-)
-
-
-class LibreOfficeChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(libreofficeconfig)
-        checkerconfig.update(openofficeconfig)
-        super().__init__(**kwargs)
-
-    @critical
-    def validxml(self, str1, str2) -> bool:
-        """
-        Check that all XML/HTML open/close tags has close/open pair in the
-        translation.
-        """
-        for location in self.locations:
-            if location.endswith((".xrm", ".xhp")):
-                opentags = []
-                match = re.search(lo_tag_re, str2)
-                while match:
-                    acttag = match.group(0)
-                    if acttag.startswith("</"):
-                        if match.group("tag") in lo_emptytags:
-                            raise FilterFailure(
-                                f"»{acttag}« should be self-closing/empty"
-                            )
-                        if len(opentags) == 0:
-                            raise FilterFailure(f"There is no open tag for »{acttag}«")
-                        opentag = opentags.pop()
-                        if tagname(acttag) != f"/{tagname(opentag)}":
-                            raise FilterFailure(
-                                f"Open tag »{opentag}« and close tag »{acttag}« "
-                                "don't match"
-                            )
-                    elif acttag.endswith("/>"):
-                        if match.group("tag") not in lo_emptytags:
-                            raise FilterFailure(
-                                f"»{acttag}« should not be self-closing/empty"
-                            )
-                    else:
-                        opentags.append(acttag)
-                    str2 = str2[match.end(0) :]
-                    match = re.search(lo_tag_re, str2)
-                if len(opentags) != 0:
-                    raise FilterFailure(f"There is no close tag for »{opentags.pop()}«")
-        return True
-
-    @critical
-    def pythonbraceformat(self, str1, str2) -> bool:
-        """Not used in LibreOffice."""
-        return True
-
-
-mozillaconfig = CheckerConfig(
-    accelmarkers=["&"],
-    varmatches=[
-        ("&", ";"),
-        ("%", "%"),
-        ("%", 1),
-        ("$", "$"),
-        ("$", None),
-        ("#", 1),
-        ("${", "}"),
-        ("$(^", ")"),
-        ("{{", "}}"),
-    ],
-    criticaltests=["accelerators"],
-)
-
-
-class MozillaChecker(StandardChecker):
-    accelerators_skipped_scripts = [
-        # spellchecker:off
-        "Deva",
-        "Beng",
-        "Tibt",
-        "Orya",
-        "Gujr",
-        "Khmr",
-        "Knda",
-        "Laoo",
-        "Mlym",
-        "Mymr",
-        "Sind",
-        "Taml",
-        "assamese",
-        "perso-arabic",
-        "mon",
-        "chinese",
-        # spellchecker:on
-    ]
-
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(mozillaconfig)
-        super().__init__(**kwargs)
-
-    @extraction
-    def credits(self, str1, str2) -> bool:
-        """
-        Checks for messages containing translation credits instead of
-        normal translations.
-
-        Some projects have consistent ways of giving credit to translators by
-        having a unit or two where translators can fill in their name and
-        possibly their contact details. This test allows you to find these
-        units easily to check that they are completed correctly and also
-        disables other tests that might incorrectly get triggered for these
-        units (such as urls, emails, etc.)
-        """
-        for location in self.locations:
-            if location in {"MOZ_LANGPACK_CONTRIBUTORS", "credit.translation"}:
-                raise FilterFailure("Don't translate. Just credit the translators.")
-
-        return True
-
-    mozilla_dialog_re = re.compile(
-        r"""(                         # option pair "key: value;"
-                                      (?P<key>[-a-z]+)           # key
-                                      :\s+                       # separator
-                                      (?P<number>\d+(?:[.]\d+)?) # number
-                                      (?P<unit>[a-z][a-z]);?     # units
-                                      )+                         # multiple pairs
-                                   """,
-        re.VERBOSE,
-    )
-    mozilla_dialog_valid_units = ["em", "px", "ch"]
-
-    @critical
-    def dialogsizes(self, str1, str2) -> bool:
-        """
-        Checks that dialog sizes are not translated.
-
-        This is a Mozilla specific test. Mozilla uses a language called XUL to
-        define dialogues and screens. This can make use of CSS to specify
-        properties of the dialogue. These properties include things such as the
-        width and height of the box. The size might need to be changed if the
-        dialogue size changes due to longer translations. Thus translators can
-        change these settings. But you are only meant to change the number not
-        translate the words 'width' or 'height'. This check capture instances
-        where these are translated. It will also catch other types of errors in
-        these units.
-        """
-        # Example: "width: 635px; height: 400px;"
-        if "width" in str1 or "height" in str1:
-            str1pairs = self.mozilla_dialog_re.findall(str1)
-
-            if str1pairs:
-                str2pairs = self.mozilla_dialog_re.findall(str2)
-
-                if len(str1pairs) != len(str2pairs):
-                    raise FilterFailure("A dialog pair is missing")
-
-                for i, pair1 in enumerate(str1pairs):
-                    pair2 = str2pairs[i]
-
-                    if pair1[0] != pair2[0]:  # Only check pairs that differ
-                        if len(pair2) != 4:
-                            raise FilterFailure("A part of the dialog pair is missing")
-
-                        if pair1[1] not in pair2:  # key
-                            raise FilterFailure(
-                                f"Do not translate the key '{pair1[1]}'"
-                            )
-
-                        # FIXME we could check more carefully for numbers in pair1[2]
-                        if pair2[3] not in self.mozilla_dialog_valid_units:
-                            raise FilterFailure(
-                                f"Units should be one of '{', '.join(self.mozilla_dialog_valid_units)}'. The source string uses '{pair1[3]}'"
-                            )
-
-        return True
-
-    @functional
-    def numbers(self, str1, str2):
-        """
-        Checks that numbers are not translated.
-
-        Special handling for Mozilla to ignore entries that are dialog sizes.
-        """
-        if self.mozilla_dialog_re.findall(str1):
-            return True
-
-        return super().numbers(str1, str2)
-
-    @functional
-    def unchanged(self, str1, str2):
-        """
-        Checks whether a translation is basically identical to the original
-        string.
-
-        Special handling for Mozilla to ignore entries that are dialog sizes.
-        """
-        if (
-            self.mozilla_dialog_re.findall(str1)
-            or str1.strip().lstrip(string.digits) in self.mozilla_dialog_valid_units
-        ):
-            return True
-
-        return super().unchanged(str1, str2)
-
-    @cosmetic
-    def accelerators(self, str1, str2):
-        """
-        Checks whether accelerators are consistent between the
-        two strings.
-
-        For Mozilla we lower the severity to cosmetic, and for some languages
-        it also ensures accelerators are absent in the target string since some
-        languages do not use accelerators, for example Indic languages.
-        """
-        # Mozilla's specific no-accelerators behavior.
-        if self.config.language_script in self.accelerators_skipped_scripts:
-            str2 = self.filtervariables(str2)
-            messages = []
-
-            for accelmarker in self.config.accelmarkers:
-                counter2 = decoration.countaccelerators(
-                    accelmarker,
-                    self.config.lang.validaccel,
-                )
-                if counter2(str2)[0] > 0:
-                    messages.append(
-                        f"Accelerator '{accelmarker}' should not appear in translation"
-                    )
-
-            if messages:
-                raise FilterFailure(messages)
-
-            return True
-
-        # Default accelerators behavior.
-        return super().accelerators(str1, str2)
-
-
-drupalconfig = CheckerConfig(
-    varmatches=[("%", None), ("@", None), ("!", None)],
-)
-
-
-class DrupalChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(drupalconfig)
-        super().__init__(**kwargs)
-
-
-gnomeconfig = CheckerConfig(
-    accelmarkers=["_"],
-    varmatches=[("%", 1), ("$(", ")")],
-    credit_sources=["translator-credits"],
-)
-
-
-class GnomeChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(gnomeconfig)
-        super().__init__(**kwargs)
-
-    @functional
-    def gconf(self, str1, str2) -> bool:
-        """
-        Checks if we have any gconf config settings translated.
-
-        Gconf settings should not be translated so this check checks that gconf
-        settings such as "name" or "modification_date" are not translated in
-        the translation. It allows you to change the surrounding quotes but
-        will ensure that the setting values remain untranslated.
-        """
-        for location in self.locations:
-            if (
-                location.find("schemas.in") != -1
-                or location.find("gschema.xml.in") != -1
-            ):
-                gconf_attributes = gconf_attribute_re.findall(str1)
-                # stopwords = [word for word in words1 if word in self.config.notranslatewords and word not in words2]
-                stopwords = [
-                    word for word in gconf_attributes if word[1:-1] not in str2
-                ]
-
-                if stopwords:
-                    raise FilterFailure(
-                        f"Do not translate GConf attributes: {', '.join(stopwords)}"
-                    )
-
-                return True
-
-        return True
-
-
-kdeconfig = CheckerConfig(
-    accelmarkers=["&"],
-    varmatches=[("%", 1)],
-    credit_sources=["Your names", "Your emails", "ROLES_OF_TRANSLATORS"],
-)
-
-
-class KdeChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        # TODO allow setup of KDE plural and translator comments so that they do
-        # not create false positives
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(kdeconfig)
-        super().__init__(**kwargs)
-
-
-cclicenseconfig = CheckerConfig(varmatches=[("@", "@")])
-
-
-class CCLicenseChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(cclicenseconfig)
-        super().__init__(**kwargs)
-
-
-minimalconfig = CheckerConfig()
-
-
-class MinimalChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        limitfilters = kwargs.get("limitfilters")
-
-        if limitfilters is None:
-            limitfilters = ["untranslated", "unchanged", "blank"]
-            kwargs["limitfilters"] = limitfilters
-
-        checkerconfig.update(minimalconfig)
-        super().__init__(**kwargs)
-
-
-reducedconfig = CheckerConfig()
-
-
-class ReducedChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        limitfilters = kwargs.get("limitfilters")
-
-        if limitfilters is None:
-            limitfilters = [
-                "untranslated",
-                "unchanged",
-                "blank",
-                "doublespacing",
-                "doublewords",
-                "spellcheck",
-            ]
-            kwargs["limitfilters"] = limitfilters
-
-        checkerconfig.update(minimalconfig)
-        super().__init__(**kwargs)
-
-
-termconfig = CheckerConfig()
-
-
-class TermChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(termconfig)
-        super().__init__(**kwargs)
-
-
-class L20nChecker(MozillaChecker):
-    excluded_filters_for_complex_units = [
-        "escapes",
-        "newlines",
-        "tabs",
-        "singlequoting",
-        "doublequoting",
-        "doublespacing",
-        "brackets",
-        "pythonbraceformat",
-        "sentencecount",
-        "variables",
-    ]
-    complex_unit_pattern = "->"
-
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        super().__init__(**kwargs)
-
-    def run_filters(self, unit, categorised=False):
-        is_unit_complex = (
-            self.complex_unit_pattern in unit.source
-            or self.complex_unit_pattern in unit.target
-        )
-
-        saved_default_filters = {}
-        if is_unit_complex:
-            saved_default_filters = self.defaultfilters
-            self.defaultfilters = {
-                key: value
-                for (key, value) in self.defaultfilters.items()
-                if key not in self.excluded_filters_for_complex_units
-            }
-
-        result = super().run_filters(unit, categorised=categorised)
-
-        if is_unit_complex:
-            self.defaultfilters = saved_default_filters
-
-        return result
-
-
-iosconfig = CheckerConfig(
-    varmatches=[("$(", ")"), ("%", "@")],
-)
-
-
-class IOSChecker(StandardChecker):
-    def __init__(self, **kwargs) -> None:
-        checkerconfig = kwargs.get("checkerconfig")
-
-        if checkerconfig is None:
-            checkerconfig = CheckerConfig()
-            kwargs["checkerconfig"] = checkerconfig
-
-        checkerconfig.update(iosconfig)
-        super().__init__(**kwargs)
-
-
-projectcheckers = {
-    "minimal": MinimalChecker,
-    "standard": StandardChecker,
-    "reduced": ReducedChecker,
-    "openoffice": OpenOfficeChecker,
-    "libreoffice": LibreOfficeChecker,
-    "mozilla": MozillaChecker,
-    "kde": KdeChecker,
-    "wx": KdeChecker,
-    "gnome": GnomeChecker,
-    "creativecommons": CCLicenseChecker,
-    "drupal": DrupalChecker,
-    "terminology": TermChecker,
-    "ios": IOSChecker,
-}
-
-
-class StandardUnitChecker(UnitChecker):
-    """The standard checks for common checks on translation units."""
-
-    @extraction
-    def isfuzzy(self, unit) -> bool:
-        """
-        Check if the unit has been marked fuzzy.
-
-        If a message is marked fuzzy in the PO file then it is extracted.
-        Note this is different from ``--fuzzy`` and ``--nofuzzy`` options which
-        specify whether tests should be performed against messages marked
-        fuzzy.
-        """
-        return not unit.isfuzzy()
-
-    @extraction
-    def isreview(self, unit) -> bool:
-        """
-        Check if the unit has been marked review.
-
-        If you have made use of the 'review' flags in your translations::
-
-          # (review) reason for review
-          # (pofilter) testname: explanation for translator
-
-        Then if a message is marked for review in the PO file it will be
-        extracted. Note this is different from ``--review`` and ``--noreview``
-        options which specify whether tests should be performed against
-        messages already marked as under review.
-        """
-        return not unit.isreview()
-
-    @critical
-    def nplurals(self, unit):
-        """
-        Checks for the correct number of noun forms for plural translations.
-
-        This uses the plural information in the language module of the
-        Translate Toolkit. This is the same as the Gettext nplural value. It
-        will check that the number of plurals required is the same as the
-        number supplied in your translation.
-        """
-        if unit.hasplural():
-            # if we don't have a valid nplurals value, don't run the test
-            nplurals = self.config.lang.nplurals
-
-            if nplurals > 0:
-                return len(list(filter(None, unit.target.strings))) == nplurals
-
-        return True
-
-    @extraction
-    def hassuggestion(self, unit) -> bool:
-        """
-        Checks if there is at least one suggested translation for this unit.
-
-        If a message has a suggestion (an alternate translation stored in
-        alt-trans units in XLIFF and .pending files in PO) then these will be
-        extracted. This is used by Pootle and is probably only useful in
-        pofilter when using XLIFF files.
-        """
-        self.suggestion_store = getattr(self, "suggestion_store", None)
-        suggestions = []
-
-        if self.suggestion_store:
-            suggestions = self.suggestion_store.findunits(unit.source)
-        elif getattr(unit, "getalttrans", None):
-            # TODO: we probably want to filter them somehow
-            suggestions = unit.getalttrans()
-
-        return not bool(suggestions)
-
-
-# TODO: convert these to proper unit tests
-def runtests(str1, str2, ignorelist=()):
-    """Verifies that the tests pass for a pair of strings."""
-    str1 = data.normalize(str1)
-    str2 = data.normalize(str2)
-    unit = base.TranslationUnit(str1)
-    unit.target = str2
-    checker = StandardChecker(excludefilters=ignorelist)
-    failures = checker.run_filters(unit)
-
-    for test, value in failures.items():
-        print(  # ruff:ignore[print]
-            f"failure: {test}: {value['message']}\n  {str1!r}\n  {str2!r}"
-        )
-
-    return failures
-
-
-def batchruntests(pairs) -> None:
-    """Runs test on a batch of string pairs."""
-    passed, numpairs = 0, len(pairs)
-
-    for str1, str2 in pairs:
-        if runtests(str1, str2):
-            passed += 1
-
-    print(f"\ntotal: {passed}/{numpairs} pairs passed")  # ruff:ignore[print]
-
-
-if __name__ == "__main__":
-    testset = [
-        (r"simple", r"somple"),
-        (r"\this equals \that", r"does \this equal \that?"),
-        (r"this \'equals\' that", r"this 'equals' that"),
-        (r" start and end! they must match.", r"start and end! they must match."),
-        (
-            r"check for matching %variables marked like %this",
-            r"%this %variable is marked",
-        ),
-        (
-            r"check for mismatching %variables marked like %this",
-            r"%that %variable is marked",
-        ),
-        (r"check for mismatching %variables% too", r"how many %variable% are marked"),
-        (r"%% %%", r"%%"),
-        (r"Row: %1, Column: %2", r"Mothalo: %1, Kholomo: %2"),
-        (r"simple lowercase", r"it is all lowercase"),
-        (r"simple lowercase", r"It Is All Lowercase"),
-        (r"Simple First Letter Capitals", r"First Letters"),
-        (r"SIMPLE CAPITALS", r"First Letters"),
-        (r"SIMPLE CAPITALS", r"ALL CAPITALS"),
-        (r"forgot to translate", r"  "),
-    ]
-    batchruntests(testset)
